[watchman] use -Werror, fix all warnings
Summary:
depends on D645597.  Enable more pedantic warnings,
and because I've seen this bite OSS projects in the ass when
someone uses a different gcc version, add a configure argument
"--enable-lenient" to relax the warnings.

Refactor the tap library build and make it an actual library
in anticipation of adding more tests; will save us from rebuilding
the object file multiple times.

Test Plan: `make integration`

Reviewers: jakubv, keegancsmith, zoel

Reviewed By: keegancsmith

CC: fugalh, scottg

Differential Revision: https://phabricator.fb.com/D645610

+++ ht.c
@@ -301,7 +301,7 @@ bool w_ht_next(w_ht_t *ht, w_ht_iter_t *iter)
 {
   struct watchman_hash_bucket *b = iter->ptr;
 
-  if (iter->slot != -1 && iter->slot >= ht->table_size) {
+  if (iter->slot != (uint32_t)-1 && iter->slot >= ht->table_size) {
     return false;
   }
